requirements.txt
census==0.8.15
CensusData==1.10
certifi==2020.11.8
chardet==3.0.4
future==0.18.2
idna==2.10
jellyfish==0.6.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
six==1.15.0
urllib3==1.25.11
us==2.0.2
